+++ tests/frontend/org/voltdb/benchmark/overhead/procedures/measureOverheadMultipartition42Strings.java
@@ -35,7 +35,7 @@ public class measureOverheadMultipartition42Strings extends VoltProcedure {
 
     public final SQLStmt getID = new SQLStmt("SELECT NO_O_ITEM FROM NEWORDER WHERE NO_O_ID = ?;");
 
-    public VoltTable[] run(long no_o_id, String arg2, String arg3, String arg4,String arg5,
+    public VoltTable[] run(int no_o_id, String arg2, String arg3, String arg4,String arg5,
             String arg6, String arg7, String arg8, String arg9, String arg10,
             String arg11, String arg12, String arg13, String arg14, String arg15,
             String arg16, String arg17, String arg18, String arg19, String arg20,
